Get the names of the facts of a star database

Obtain the names of the facts of a star database.

Usage

get_fact_names(db)

# S3 method for class 'star_database'
get_fact_names(db)

Arguments

db

A star_database object.

Value

A vector of strings, fact names.

See also

Examples

names <- star_database(mrs_cause_schema, ft_num) |>
  get_fact_names()