Get new dimension instances

From the planned update, it obtains the instances of the update dimensions that are not included in the star database dimensions to be updated.

Usage

get_new_dimension_instances(sdbu)

# S3 method for class 'star_database_update'
get_new_dimension_instances(sdbu)

Arguments

sdbu

A star_database_update object.

Value

A list of tibble objects.

See also

Examples

f1 <-
  flat_table('ft_num', ft_cause_rpd[ft_cause_rpd$City != 'Cambridge' &
                                      ft_cause_rpd$WEEK != '4',]) |>
  as_star_database(mrs_cause_schema_rpd) |>
  role_playing_dimension(rpd = "When",
                         roles = c("When Available", "When Received"))
f2 <- flat_table('ft_num2', ft_cause_rpd[ft_cause_rpd$City != 'Bridgeport' &
                                           ft_cause_rpd$WEEK != '2',])
f2 <- f2 |>
  update_according_to(f1)
dim_instances <- f2 |>
  get_new_dimension_instances()