star_schema S3 class

An empty star_schema object is created in which definition of facts and dimensions can be added.

Usage

star_schema()

Value

A star_schema object.

Details

To get a star database (a star_database object) we need a flat table and a star_schema object. The definition of facts and dimensions in the star_schema object is made from the flat table columns.

See also

Examples

s <- star_schema()